feat(system): 新增系统会话管理功能
- 创建SysSession实体类用于存储用户会话信息
- 实现SysSessionController提供会话列表查询和踢出会话接口
- 定义ISysSessionService服务接口和SysSessionServiceImpl实现类
- 添加SysSessionQueryDTO查询数据传输对象支持条件过滤
- 集成分页查询功能支持会话列表分页展示
- 实现基于用户名、昵称、IP、平台类型的会话过滤功能
- 集成SessionHelper实现会话踢出功能

+++ junoyi-module/junoyi-module-system/src/main/java/com/junoyi/system/service/SysSessionServiceImpl.java
@@ -46,7 +46,7 @@ public class SysSessionServiceImpl implements ISysSessionService {
                 ? filteredList.subList(fromIndex, toIndex) 
                 : List.of();
         
-        return PageResult.of(pageList, total, pageNum, pageSize);
+        return PageResult.of(pageList, (long) total, pageNum, pageSize);
     }
 
     /**
